requirements.txt
numpy
tqdm
cv2
pyyaml

# Ray Casting
trimesh

# Depth & Image Diffusion
transformers
torch

# UVP
torchvision